Fit prolong Model

prolong() is designed to take an \(n \times t\) outcome matrix and an \(n \times p \times t\) array of covariates and automatically run the entire processing and model fitting process, returning a list of coefficients with corresponding variable names.

Usage

prolong(
  Y,
  X,
  lambda1 = NULL,
  lambda2 = NULL,
  lambdar = NULL,
  groups = TRUE,
  foldids = NULL,
  optimvals = c(1, 0.01)
)

Arguments

Y

Input response matrix, with n rows and t columns

X

Input covariate array, with n rows, p columns, and t slices

lambda1

Lasso/group lasso parameter, if left NULL this parameter will be chosen via a cross-validation that keeps each subject's observations across time points together. It is recommended to save the lambda2 and lambdar values from the first run of prolong for future runs, since the optimization step takes the longest

lambda2

Laplacian penalty parameter, if left NULL will be chosen along with lambdar via MLE, see supplementary material in (Steiner et al. 2022)

lambdar

Nuisance parameter, if left NULL will be chosen along with lambdar via MLE. Added to the the diagonal elements of the laplacian matrix to get the invertibility required for the MLE

groups

Optional pre-specified groups. If NULL or FALSE, lasso will be used. If left as TRUE, then there will be p groups each containing the observations across time points

foldids

Optional pre-specified foldids for the cv. Should be of length n. If left NULL, subjects will be automatically split into 5 folds

optimvals

Initial values of lambda2 and lambdar to be used by optim() for the MLE of lambda2 and lambdar

Value

A named numeric vector of coefficients from the lasso or group lasso

Details

First, the data is reshaped into the first-differenced vectorized \(n(t-1)\) length Y and first-differenced matricized \(n(t-1) \times pt(t-1)/2\) X with a corresponding dependence matrix. Next, hyperparameters for the graph laplacian-based network penalty are found via MLE and the parameter for lasso/group lasso is found via a careful implementation of cross-validation. Lastly, a group lasso + laplacian or lasso + laplacian model is implemented, and its bias-adjusted coefficients are returned.

References

Steiner A, Abbas K, Brzyski D, Pączek K, Randolph TW, Goñi J, Harezlak J (2022). “Incorporation of spatial- and connectivity-based cortical brain region information in regularized regression: Application to Human Connectome Project data.” Frontiers in Neuroscience, 16. ISSN 1662-453X, doi:10.3389/fnins.2022.957282 , https://www.frontiersin.org/articles/10.3389/fnins.2022.957282. Yang Y, Zou H (2015). “A fast unified algorithm for solving group-lasso penalize learning problems.” Statistics and Computing, 25(6), 1129--1141. ISSN 1573-1375, doi:10.1007/s11222-014-9498-5 , 2022-09-06. Yuan M, Lin Y (2005). “Model Selection and Estimation in Regression with Grouped Variables.” Journal of the Royal Statistical Society Series B: Statistical Methodology, 68(1), 49-67. ISSN 1369-7412, doi:10.1111/j.1467-9868.2005.00532.x , https://academic.oup.com/jrsssb/article-pdf/68/1/49/49794691/jrsssb\_68\_1\_49.pdf.

Examples

if (FALSE) {
coefs <- prolong(Ymatrix, Xarray)
coefs

coefs <- prolong(Ymatrix, Xarray, lambda2 = .001, lambdar = 10, groups = )
coefs
}